Implemented Ground Truth Assessment Tool
This tool allows a video to be analyzed frame-by-frame and for the Overwatch Hero and weapon action to be recorded. It outputs a comma seperated file that can be used to identify and extract desired frames, or assess the accuracy of an object detection method by comparing against the ground truth.